Move `MediaChoiceWidget` where it belongs; remove unused `MediaChoiceField`
